Algorithms/src/main/java/com/ankur/interview/tries/Trie.java
package com.ankur.interview.tries;

public class Trie {
    private TrieNode root;

    public Trie() {
        root = new TrieNode();
    }

    public void insert(String word) {
        TrieNode current = root;
        for (int i = 0; i < word.length(); i++) {
            char ch = word.charAt(i);
            TrieNode node = current.getChildren().get(ch);
            if (node == null) {
                node = new TrieNode();
                current.getChildren().put(ch, node);
            }
            current = node;
        }
        //mark the current nodes endOfWord as true
        current.setLeaf(true);
    }

    public boolean search(String word) {
        TrieNode current = root;
        for (int i = 0; i < word.length(); i++) {
            char ch = word.charAt(i);
            TrieNode node = current.getChildren().get(ch);
            //if node does not exist for given char then return false
            if (node == null) {
                return false;
            }
            current = node;
        }
        //return true of current's endOfWord is true else return false.
        return current.isLeaf();
    }

    public boolean delete(String word) {
        return delete(root, word, 0);
    }

    private boolean delete(TrieNode current, String word, int index) {
        if (index == word.length()) {
            //when end of word is reached only delete if currrent.endOfWord is true.
            if (!current.isLeaf()) {
                return false;
            }
            current.setLeaf(false);
            //if current has no other mapping then return true
            return current.getChildren().size() == 0;
        }
        char ch = word.charAt(index);
        TrieNode node = current.getChildren().get(ch);
        if (node == null) {
            return false;
        }
        boolean shouldDeleteCurrentNode = delete(node, word, index + 1);
        //if true is returned then delete the mapping of character and trienode reference from map.
        if (shouldDeleteCurrentNode) {
            current.getChildren().remove(ch);
            //return true if no mappings are left in the map.
            return current.getChildren().size() == 0;
        }
        return false;
    }

}
